Starting Point, Refactored Engine
Engine from game Fedora Frisbee is refactored, more seperation between Engine files and Game-specfic files. Can be used to create more proejcts easier. Not sure if this repository will be a new game, or just improving the engine yet.
I do plan on making a new game, just not sure if it is in this repo.